Tournament class for collecting data
I am developing the API necessary to pull raw data from the WCF results page. This will primarily be guided by what I need it to do, which is pull full tournament results and identifiers and not much else. In that regard, the "official" WCF API would be better for getting player data, which I am ignoring.