rewrote `parsing_trial.py` using `wcf`
To make use of my new API, I rewrote the work I did in `parsing_trial.py` to do the same thing, but with `wcf` handling all of the game loading and parsing. The game "meta-data" or "macro-strategy" is still handled in much the same way, except where the `BoxScore` objects help out.

I also changed how LSFE was stored within the BoxScore object, since it was inconsistent with how I wa storing the winner of the game. Either both are boolean lists or both are indexes (0 or 1) for the team that matches the item; I chose the latter.

I also noticed that I wasn't testing for LSFE, so I added in that test.