Working on official API access
Now that I have legitimate access to the WCF database, this module will be for accessing the information through official means. I played around with the API briefly, mostly following the quick-start guide.

There are quite a few different things you can do, past just pulling information, so I'll explore and implement those as need be.

This is just a start.