testing saving data and parsing team member data
One thing I need to change in the `wcf.py` API package is the parsing data. I should keep team (and individual player?) accuracies as an additional feature or features, but the way I have it set up now is that the `wcf.Tournament` structure throws out everything that isn't the box score.

That's not good for reusability or reparsing a game or something. I need to clean up that entire module anyway, especially when I look at how nice the `wcfapi.py` module looks...